+++ CHANGELOG.md
@@ -15,8 +15,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - The RTK extension now exposes methods that allow other extensions to retrieve
   the status of RTK corrections being received from the base station.
 
-- RC simulation with UDP packets now has a configurable timeout after which the
-  RC connection is assumed to be lost.
+- MAVLink drones will now receive `RC_CHANNELS_OVERRIDE` messages when the
+  server receives RC channel change events from a simulated RC via UDP packets.
+  You need to enable the `rc_udp` extension to use this feature.
+
+- RC simulation with UDP packets in the `rc_udp` extension now has a configurable
+  timeout after which the RC connection is assumed to be lost.
 
 ## [2.7.1] - 2022-12-25